Null check added for inPSI
Fixed issue with the hunterpawn type bounty hunters. When a bounty hunter spawns that is of the forcibly spawning a pawn on your ship type it spawns a default bandit drone ship named bounty hunter. This ship has a null inPSI which causes the ship respawner function that keeps ships around that are abandoned when you leave and come back to a sector to throw an error. This prevents the bounty hunter from being spawned in correctly which leads to an infinite number of the default bandit drone ships spawning (only one every 30 seconds) but no hunter pawn.